pelkkää vihreetä

+++ Moduuli1/Suorakulmion kanta ja korkeus - piiri ja ala.py
@@ -1,5 +1,5 @@
-b = int(input("Anna suorakulmion kannan pituus senttimetreinä: "))
-h = int(input("Anna suorakulmion korkeus senttimetreinä: "))
+b = float(input("Anna suorakulmion kannan pituus senttimetreinä: "))
+h = float(input("Anna suorakulmion korkeus senttimetreinä: "))
 
 area = b * h
 perim = 2*(b + h)
